+++ engine-standalone-storage/src/sync/types.rs
@@ -1,5 +1,5 @@
 use aurora_engine::parameters;
-use aurora_engine::transaction::EthTransactionKind;
+use aurora_engine_transactions::EthTransactionKind;
 use aurora_engine_types::account_id::AccountId;
 use aurora_engine_types::H256;
